[Feat] 라벨 페이지 delete 이벤트 추가

+++ client/src/components/labels/List/Label/index.js
@@ -7,39 +7,43 @@ import {
 import ColorHandler from '@utils/colorHandler'
 import S from './style';
 
-function Label({label}) {
-  const [ isEditState, setIsEditState ] = useState(false);
+function Label({ label }) {
+  const [isEditState, setIsEditState] = useState(false);
   const dispatch = useLabelDispatch();
-  const handleEditLabel =()=>{
+  const handleEditLabel = () => {
     setIsEditState(!isEditState);
   }
-  const handleDeleteLabel =()=>{
-    deleteLabel(dispatch, label.id);
+  const handleDeleteLabel = () => {
+    if (confirm('Are you sure? Deleting a label will remove it from all issues and pull requests.')) {
+      deleteLabel(dispatch, label.id);
+    } else {
+      return false;
+    }
   }
   return (
-   <>
-   {isEditState 
-   ?
-   <NewLabel 
-    setIsEditState={setIsEditState}
-    label={label}/>
-    :
-    <S.LabelWrapper>
-      <S.LabelListWrapper>
-        <S.LabelItemWrapper>
-          <S.LabelItem fontColor={ColorHandler.getContrastColor(label.color)} color={label.color}>{label.name}</S.LabelItem>
-        </S.LabelItemWrapper>
-        <S.DescWrapper>{label.desc}</S.DescWrapper>
-        <S.ButtonsWrapper>
-          <S.EditButton
-            className="label-button-item" 
-            onClick={handleEditLabel}>Edit</S.EditButton>
-          <S.DeleteButton 
-          className="label-button-item" onClick={handleDeleteLabel}>Delete</S.DeleteButton>
-        </S.ButtonsWrapper>
-      </S.LabelListWrapper>
-    </S.LabelWrapper>
-    }
+    <>
+      {isEditState
+        ?
+        <NewLabel
+          setIsEditState={setIsEditState}
+          label={label} />
+        :
+        <S.LabelWrapper>
+          <S.LabelListWrapper>
+            <S.LabelItemWrapper>
+              <S.LabelItem fontColor={ColorHandler.getContrastColor(label.color)} color={label.color}>{label.name}</S.LabelItem>
+            </S.LabelItemWrapper>
+            <S.DescWrapper>{label.desc}</S.DescWrapper>
+            <S.ButtonsWrapper>
+              <S.EditButton
+                className="label-button-item"
+                onClick={handleEditLabel}>Edit</S.EditButton>
+              <S.DeleteButton
+                className="label-button-item" onClick={handleDeleteLabel}>Delete</S.DeleteButton>
+            </S.ButtonsWrapper>
+          </S.LabelListWrapper>
+        </S.LabelWrapper>
+      }
     </>
   );
 }
